added a new exectuor file and improved caching
